requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.54.3
tenacity==9.0.0
pybreaker==1.2.0
sentence-transformers==3.2.1
torch==2.4.1
numpy==1.26.4
insightface==0.7.3
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
pillow==10.4.0
pytest==8.3.3
httpx==0.27.2
